+++ hercules-ci-api-core/src/Hercules/API/Servant.hs
@@ -1,5 +1,21 @@
+{-# LANGUAGE PolyKinds #-}
+{-# LANGUAGE TypeFamilies #-}
+
 -- | Extras for working with servant
-module Hercules.API.Servant where
+module Hercules.API.Servant
+  ( noContent,
+
+    -- * 'Generic'
+    useApi,
+    useApiE,
+    enterApi,
+    enterApiE,
+
+    -- * Substitution
+    Substitute,
+    Placeholder,
+  )
+where
 
 import Control.Monad (void)
 import Servant.API
@@ -12,11 +28,40 @@ import Prelude
 --
 -- Ideally, this functionality would be built into a new combinator.
 useApi ::
-  (GenericServant f mode, GenericServant g mode) =>
-  (f mode -> ToServant g mode) ->
-  f mode ->
-  g mode
+  forall subapi api mode.
+  (GenericServant api mode, GenericServant subapi mode) =>
+  (api mode -> ToServant subapi mode) ->
+  api mode ->
+  subapi mode
 useApi = (Servant.API.Generic.fromServant .)
+
+-- | Like 'useApi' but constrains the @auth@ type variable that's passed to
+-- subapis.
+useApiE ::
+  forall subapi api mode a.
+  (GenericServant (api a) mode, GenericServant (subapi a) mode) =>
+  (api a mode -> ToServant (subapi a) mode) ->
+  api a mode ->
+  subapi a mode
+useApiE = useApi
+
+-- | @flip 'useApi'
+enterApi ::
+  forall subapi api mode.
+  (GenericServant api mode, GenericServant subapi mode) =>
+  api mode ->
+  (api mode -> ToServant subapi mode) ->
+  subapi mode
+enterApi = flip useApi
+
+-- | @flip 'useApiE'
+enterApiE ::
+  forall subapi api mode a.
+  (GenericServant (api a) mode, GenericServant (subapi a) mode) =>
+  api a mode ->
+  (api a mode -> ToServant (subapi a) mode) ->
+  subapi a mode
+enterApiE = flip useApi
 
 -- | 'Control.Monad.void' specialised to 'NoContent' to soothe the
 -- compiler that rightfully warns about throwing away a do notation
@@ -24,3 +69,15 @@ useApi = (Servant.API.Generic.fromServant .)
 -- if the result type changes in the future. (We'll get an error)
 noContent :: Functor m => m Servant.API.NoContent -> m ()
 noContent = void
+
+-- | A reference to the @subapi@ parameter in 'Substitute'
+data Placeholder
+
+-- | Replaces 'Placeholder' by @subapi@ in the API @target@.
+type family Substitute (target :: k) subapi :: *
+
+type instance Substitute Placeholder subapi = subapi
+
+type instance Substitute (a :> b) subapi = a :> Substitute b subapi
+
+type instance Substitute (a :<|> b) subapi = Substitute a subapi :<|> Substitute b subapi