+++ hercules-ci-api-core/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [0.1.3.0] - Unreleased
+
+### Added
+
+- Utilities for servant's `Generic` feature
+  - `Substitute`
+  - `useApiE`, `enterApi`, `enterApiE`
+
 ## [0.1.2.0] - 2020-03-07
 
 - Use `uuid` type in swagger